requirements.txt
imbox==0.9.8
yagmail==0.15.293
python-dotenv==1.0.0
openai==1.12.0 